change is super admin field to role field and add user role enum

+++ app/Http/Controllers/UserController.php
@@ -6,11 +6,13 @@ use App\ActionContracts\Users\CreateUserActionInterface;
 use App\ActionContracts\Users\EditUserActionInterface;
 use App\DataTransferObjects\Users\CreateUserDTO;
 use App\DataTransferObjects\Users\EditUserDTO;
+use App\Enums\UserRole;
 use App\Http\Requests\Users\UserStoreRequest;
 use App\Http\Requests\Users\UserUpdateRequest;
 use App\Models\User;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Support\Facades\Cache;
+use Illuminate\Validation\ValidationException;
 use Inertia\Inertia;
 use Inertia\Response;
 
@@ -56,8 +58,15 @@ class UserController extends Controller
         return redirect()->route('users.index')->with('success', 'User updated successfully.');
     }
 
+    /**
+     * @throws ValidationException
+     */
     public function destroy(User $user): RedirectResponse
     {
+        if ($user->role == UserRole::SUPER_ADMIN) {
+            return redirect()->route('users.index')->withErrors(['role' => 'You can not delete super admin.']);
+        }
+
         $user->delete();
 
         return redirect()->route('users.index')->with('success', 'User deleted successfully.');
